Fixed DT stringstream stuff + missing typecast in string.c

+++ base/string.c
@@ -8,7 +8,7 @@
 // Unicode codepoint
 fn Codepoint decodeUTF8(u8 *glyph_start) {
   Codepoint res = {0};
-  
+
   if ((*glyph_start & 0x80) == 0) {
     res.codepoint = *glyph_start;
     res.size = 1;
@@ -30,13 +30,13 @@ fn Codepoint decodeUTF8(u8 *glyph_start) {
   } else {
     Assert(false);
   }
-  
+
   return res;
 }
 
 fn Codepoint decodeUTF16(u16 *glyph_start) {
   Codepoint res = {0};
-  
+
   if (glyph_start[0] <= 0xD7FF ||
       (glyph_start[0] >= 0xE000 && glyph_start[0] <= 0xFFFF)) {
     res.size = 1;
@@ -49,7 +49,7 @@ fn Codepoint decodeUTF16(u16 *glyph_start) {
   } else {
     Assert(false);
   }
-  
+
   return res;
 }
 
@@ -115,14 +115,14 @@ fn String8 str8FromStream(Arena *arena, StringStream stream) {
     memCopy(ptr, curr->value.str, curr->value.size);
     ptr += curr->value.size;
   }
-  
+
   return str8(str, stream.total_size);
 }
 
 fn void stringstreamAppend(Arena *arena, StringStream *strlist, String8 other) {
   Assert(arena);
   Assert(strlist);
-  
+
   strlist->node_count += 1;
   strlist->total_size += other.size;
   StringNode *str = New(arena, StringNode);
@@ -162,13 +162,13 @@ fn bool strEq(String8 s1, String8 s2) {
   if (s1.size != s2.size) {
     return false;
   }
-  
+
   for (isize i = 0; i < s1.size; ++i) {
     if (s1.str[i] != s2.str[i]) {
       return false;
     }
   }
-  
+
   return true;
 }
 
@@ -178,14 +178,14 @@ fn bool strEqCstr(String8 s, const char *cstr) {
   } else if (!cstr || s.size == 0) {
     return false;
   }
-  
+
   isize i = 0;
   for (; i < s.size; ++i) {
     if (s.str[i] != cstr[i]) {
       return false;
     }
   }
-  
+
   if (cstr[i]) {
     return false;
   } else {
@@ -200,14 +200,14 @@ fn bool cstrEq(char *s1, char *s2) {
   if (!s1 || !s2) {
     return false;
   }
-  
+
   char *it1 = s1, *it2 = s2;
   for (; *it1 && *it2; ++it1, ++it2) {
     if (*it1 != *it2) {
       return false;
     }
   }
-  
+
   return !*it1 && !*it2;
 }
 
@@ -216,13 +216,13 @@ fn bool strIsSignedInteger(String8 s) {
   if (*curr == '-' || *curr == '+') {
     ++curr;
   }
-  
+
   for (; curr < s.str + s.size; ++curr) {
     if (!charIsDigit(*curr)) {
       return false;
     }
   }
-  
+
   return true;
 }
 
@@ -232,7 +232,7 @@ fn bool strIsInteger(String8 s) {
       return false;
     }
   }
-  
+
   return true;
 }
 
@@ -242,7 +242,7 @@ fn bool strIsFloating(String8 s) {
   if (*curr == '-' || *curr == '+') {
     ++curr;
   }
-  
+
   for (; curr < s.str + s.size; ++curr) {
     if (!charIsDigit(*curr)) {
       if (*curr == '.' && !decimal_found) {
@@ -252,7 +252,7 @@ fn bool strIsFloating(String8 s) {
       }
     }
   }
-  
+
   return true;
 }
 
@@ -262,12 +262,12 @@ inline fn bool strIsNumerical(String8 s) {
 
 fn i64 i64FromStr(String8 s) {
   i64 res = 0, decimal = 1;
-  
+
   Assert(strIsSignedInteger(s));
   for (u8 *curr = s.str + s.size - 1; curr > s.str; --curr, decimal *= 10) {
     res += (*curr - '0') * decimal;
   }
-  
+
   if (s.str[0] == '-') {
     return -res;
   } else if (s.str[0] == '+') {
@@ -279,12 +279,12 @@ fn i64 i64FromStr(String8 s) {
 
 fn u64 u64FromStr(String8 s) {
   i64 res = 0, decimal = 1;
-  
+
   Assert(strIsInteger(s));
   for (u8 *curr = s.str + s.size - 1; curr >= s.str; --curr, decimal *= 10) {
     res += (*curr - '0') * decimal;
   }
-  
+
   return res;
 }
 
@@ -300,7 +300,7 @@ fn usize strHash(String8 s) {
   for (isize i = 0; i < s.size; ++i) {
     hash = (hash << 5) + hash + s.str[i];
   }
-  
+
   return hash;
 }
 
@@ -309,25 +309,25 @@ fn String8 stringifyI64(Arena *arena, i64 n) {
   if (n < 0) {
     n = -n;
   }
-  
+
   usize i = 0, approx = 30;
   u8 *str = New(arena, u8, approx);
   for (; n > 0; ++i, n /= 10) {
     str[i] = n % 10 + '0';
   }
-  
+
   if (sign < 0) {
     str[i++] = '-';
   }
-  
+
   for (usize j = 0, k = i - 1; j < k; ++j, --k) {
     u8 tmp = str[k];
     str[k] = str[j];
     str[j] = tmp;
   }
-  
+
   arenaPop(arena, approx - i);
-  
+
   return str8(str,i);
 }
 
@@ -337,26 +337,26 @@ fn String8 stringifyU64(Arena *arena, u64 n) {
   for (; n > 0; ++i, n /= 10) {
     str[i] = n % 10 + '0';
   }
-  
+
   for (usize j = 0, k = i - 1; j < k; ++j, --k) {
     u8 tmp = str[k];
     str[k] = str[j];
     str[j] = tmp;
   }
-  
+
   arenaPop(arena, approx - i);
-  
+
   return str8(str, i);
 }
 
 fn String8 stringifyF64(Arena *arena, f64 n) {
   usize approx = 100, size = 0;
   u8 *str = New(arena, u8, approx);
-  
+
   // TODO: maybe implement `sprintf`?
   size = sprintf((char *)str, "%f", n);
   arenaPop(arena, approx - size);
-  
+
   return str8(str, size);
 }
 
@@ -364,7 +364,7 @@ fn isize str8len(char *chars) {
   char *start = chars;
   for (; *start; ++start)
     ;
-  
+
   return start - chars;
 }
 
@@ -373,7 +373,7 @@ fn String8 strFormat(Arena *arena, const char *fmt, ...) {
   va_start(args, fmt);
   String8 res = strFormatVa(arena, fmt, args);
   va_end(args);
-  
+
   return res;
 }
 
@@ -381,12 +381,12 @@ fn String8 strFormatVa(Arena *arena, const char *fmt, va_list args) {
   va_list args2;
   va_copy(args2, args);
   u32 needed_bytes = vsnprintf(0, 0, fmt, args2) + 1;
-  
+
   String8 res;
   res.str = New(arena, u8, needed_bytes);
   res.size = vsnprintf((char *)res.str, needed_bytes, fmt, args);
   res.str[res.size] = 0;
-  
+
   va_end(args2);
   return res;
 }
@@ -417,15 +417,15 @@ fn String8 longestCommonSubstring(Arena *arena, String8 s1, String8 s2) {
     usize m;
     usize n;
   };
-  
+
   usize m = s1.size + 1;
   usize n = s2.size + 1;
-  
+
   Table c = {0};
   c.x = New(arena, usize, m*n);
   c.m = m;
   c.n = n;
-  
+
   for(usize i = 1; i < m; ++i) {
     for(usize j = 1; j < n; ++j) {
       if(s1.str[i-1] == s2.str[j-1]) {
@@ -435,7 +435,7 @@ fn String8 longestCommonSubstring(Arena *arena, String8 s1, String8 s2) {
       }
     }
   }
-  
+
   usize size = at(c,m-1,n-1);
   u8 *str = New(arena, u8, size);
   usize idx = size - 1;
@@ -457,27 +457,27 @@ fn String8 longestCommonSubstring(Arena *arena, String8 s1, String8 s2) {
 
 fn String8 upperFromStr(Arena *arena, String8 s) {
   String8 res = {New(arena, u8, s.size), s.size};
-  
+
   for (isize i = 0; i < s.size; ++i) {
     res.str[i] = charToUpper(s.str[i]);
   }
-  
+
   return res;
 }
 
 fn String8 lowerFromStr(Arena *arena, String8 s) {
   String8 res = {New(arena, u8, s.size), s.size};
-  
+
   for (isize i = 0; i < s.size; ++i) {
     res.str[i] = charToLower(s.str[i]);
   }
-  
+
   return res;
 }
 
 fn String8 capitalizeFromStr(Arena *arena, String8 s) {
   String8 res = {New(arena, u8, s.size), s.size};
-  
+
   res.str[0] = charToUpper(s.str[0]);
   for (isize i = 1; i < s.size; ++i) {
     if (charIsSpace(s.str[i])) {
@@ -488,20 +488,20 @@ fn String8 capitalizeFromStr(Arena *arena, String8 s) {
       res.str[i] = charToLower(s.str[i]);
     }
   }
-  
+
   return res;
 }
 
 fn StringStream strSplit(Arena *arena, String8 s, char ch) {
   StringStream res = {0};
-  
+
   usize prev = 0;
   for (isize i = 0; i < s.size;) {
     if (s.str[i] == ch) {
       if (prev != i) {
         stringstreamAppend(arena, &res, strRange(s, prev, i));
       }
-      
+
       do {
         prev = ++i;
       } while (s.str[i] == ch);
@@ -509,11 +509,11 @@ fn StringStream strSplit(Arena *arena, String8 s, char ch) {
       ++i;
     }
   }
-  
+
   if (prev != s.size) {
     stringstreamAppend(arena, &res, strRange(s, prev, s.size));
   }
-  
+
   return res;
 }
 
@@ -523,7 +523,7 @@ fn usize strFindFirst(String8 s, char ch) {
       return curr - s.str;
     }
   }
-  
+
   return 0;
 }
 
@@ -531,7 +531,7 @@ fn usize strFindFirstSubstr(String8 haystack, String8 needle) {
   if (haystack.size < needle.size) {
     return 0;
   }
-  
+
   for (isize i = 0; i < haystack.size; ++i) {
     if (haystack.str[i] == needle.str[0]) {
       for (isize j = 0; i < needle.size; ++j) {
@@ -539,13 +539,13 @@ fn usize strFindFirstSubstr(String8 haystack, String8 needle) {
           goto outer;
         }
       }
-      
+
       return i;
     }
-    
+
     outer: ;
   }
-  
+
   return 0;
 }
 
@@ -555,7 +555,7 @@ fn bool strContains(String8 s, char ch) {
       return true;
     }
   }
-  
+
   return false;
 }
 
@@ -586,11 +586,11 @@ fn String8 strTrim(String8 s) {
   isize start = 0;
   for (; start < s.size && (s.str[start] == ' ' || s.str[start] == '\t' ||
                             s.str[start] == '\n' || s.str[start] == '\r'); ++start);
-  
+
   isize end = s.size;
   for (; end >= 0 && (s.str[end] == ' ' || s.str[end] == '\t' ||
                       s.str[end] == '\n') || s.str[end] == '\r'; --end);
-  
+
   String8 res = {
     .str = s.str + start,
     .size = end - start,
@@ -604,13 +604,13 @@ bool str16Eq(String16 s1, String16 s2) {
   if (s1.size != s2.size) {
     return false;
   }
-  
+
   for (isize i = 0; i < s1.size; ++i) {
     if (s1.str[i] != s2.str[i]) {
       return false;
     }
   }
-  
+
   return true;
 }
 
@@ -621,7 +621,7 @@ fn usize cstring16_length(u16 *str){
 }
 
 fn String16 str16_cstr(u16 *str){
-  String16 result = { str, cstring16_length(str) };
+  String16 result = { str, (isize)cstring16_length(str) };
   return result;
 }
 
@@ -629,13 +629,13 @@ fn bool str32Eq(String32 s1, String32 s2) {
   if (s1.size != s2.size) {
     return false;
   }
-  
+
   for (isize i = 0; i < s1.size; ++i) {
     if (s1.str[i] != s2.str[i]) {
       return false;
     }
   }
-  
+
   return true;
 }
 
@@ -644,17 +644,17 @@ fn bool str32Eq(String32 s1, String32 s2) {
 fn String8 UTF8From16(Arena *arena, String16 in) {
   usize res_size = 0, approx_size = in.size * 4;
   u8 *bytes = New(arena, u8, approx_size), *res_offset = bytes;
-  
+
   Codepoint codepoint = {0};
   for (u16 *start = in.str, *end = in.str + in.size; start < end;
        start += codepoint.size) {
     codepoint = decodeUTF16(start);
-    
+
     u8 utf8_codepoint_size = encodeUTF8(res_offset, codepoint);
     res_size += utf8_codepoint_size;
     res_offset += utf8_codepoint_size;
   }
-  
+
   arenaPop(arena, (approx_size - res_size));
   return str8(bytes, res_size);
 }
@@ -662,17 +662,17 @@ fn String8 UTF8From16(Arena *arena, String16 in) {
 fn String8 UTF8From32(Arena *arena, String32 in) {
   usize res_size = 0, approx_size = in.size * 4;
   u8 *bytes = New(arena, u8, approx_size), *res_offset = bytes;
-  
+
   Codepoint codepoint = {0};
   for (u32 *start = in.str, *end = in.str + in.size; start < end;
        start += codepoint.size) {
     codepoint = decodeUTF32(start);
-    
+
     u8 utf8_codepoint_size = encodeUTF8(res_offset, codepoint);
     res_size += utf8_codepoint_size;
     res_offset += utf8_codepoint_size;
   }
-  
+
   arenaPop(arena, (approx_size - res_size));
   return str8(bytes, res_size);
 }
@@ -680,17 +680,17 @@ fn String8 UTF8From32(Arena *arena, String32 in) {
 fn String16 UTF16From8(Arena *arena, String8 in) {
   isize res_size = 0, approx_size = in.size * 2;
   u16 *words = New(arena, u16, approx_size), *res_offset = words;
-  
+
   Codepoint codepoint = {0};
   for (u8 *start = in.str, *end = in.str + in.size; start < end;
        start += codepoint.size) {
     codepoint = decodeUTF8(start);
-    
+
     u8 utf16_codepoint_size = encodeUTF16(res_offset, codepoint);
     res_size += utf16_codepoint_size;
     res_offset += utf16_codepoint_size;
   }
-  
+
   arenaPop(arena, (approx_size - res_size));
   String16 res = {words, res_size};
   return res;
@@ -699,17 +699,17 @@ fn String16 UTF16From8(Arena *arena, String8 in) {
 fn String16 UTF16From32(Arena *arena, String32 in) {
   isize res_size = 0, approx_size = in.size * 2;
   u16 *words = New(arena, u16, approx_size), *res_offset = words;
-  
+
   Codepoint codepoint = {0};
   for (u32 *start = in.str, *end = in.str + in.size; start < end;
        start += codepoint.size) {
     codepoint = decodeUTF32(start);
-    
+
     u8 utf16_codepoint_size = encodeUTF16(res_offset, codepoint);
     res_size += utf16_codepoint_size;
     res_offset += utf16_codepoint_size;
   }
-  
+
   arenaPop(arena, (approx_size - res_size));
   String16 res = {words, res_size};
   return res;
@@ -718,14 +718,14 @@ fn String16 UTF16From32(Arena *arena, String32 in) {
 fn String32 UTF32From8(Arena *arena, String8 in) {
   isize res_size = 0, approx_size = in.size * 2;
   u32 *dwords = New(arena, u32, approx_size), *res_offset = dwords;
-  
+
   Codepoint cp = {0};
   for (u8 *start = in.str, *end = in.str + in.size; start < end;
        start += cp.size, ++res_size) {
     cp = decodeUTF8(start);
     *res_offset++ = cp.codepoint;
   }
-  
+
   arenaPop(arena, (approx_size - res_size));
   String32 res = {dwords, res_size};
   return res;
@@ -734,14 +734,14 @@ fn String32 UTF32From8(Arena *arena, String8 in) {
 fn String32 UTF32From16(Arena *arena, String16 in) {
   isize res_size = 0, approx_size = in.size * 2;
   u32 *dwords = New(arena, u32, approx_size), *res_offset = dwords;
-  
+
   Codepoint cp = {0};
   for (u16 *start = in.str, *end = in.str + in.size; start < end;
        start += cp.size, ++res_size) {
     cp = decodeUTF16(start);
     *res_offset++ = cp.codepoint;
   }
-  
+
   arenaPop(arena, (approx_size - res_size));
   String32 res = {dwords, res_size};
   return res;
